fix: omp num threads, resolves #163, resolves #137
This ensure that OMP_NUM_THREADS and related environment variables are set to '1' if they are unspecified.

If they are specified externally, we emit a warning message
* about potential crashes on macOS
* about potential performance issues on other OS
but do not attempt to change the variables to allow the user to perform scaling studies and evaluate threading performance themselves.

In addition, we use threadpoolctl to limit the number of threads used by thread pools in libraries such as OpenBLAS, MKL, BLIS, Veclib, and NumExpr to 1.

Two unit tests ensure these settings survive and are applied.

Furthermore, we let every ray worker test whether parallelization has really been turned off.
If this is not the case, an error is raised to prevent oversubscription.
This is also being tested using a unit test that spawns a subprocess with OMP_NUM_THREADS set to 42.

Fixes #163
Fixes #137